# wanaku-router/wanaku-router-backend/src/main/webui/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "InputSchema" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "properties" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/components/schemas/Property"
            }
          },
          "required" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "Param" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          }
        }
      },
...
      "Property" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "target" : {
            "type" : "string"
          },
          "scope" : {
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          }
        }
      },
      "ResourcePayload" : {
        "type" : "object",
        "properties" : {
          "resourceReference" : {
            "$ref" : "#/components/schemas/ResourceReference"
          },
          "configurationData" : {
            "type" : "string"
          },
          "secretsData" : {
            "type" : "string"
          },
          "payload" : {
            "$ref" : "#/components/schemas/ResourceReference"
          }
        }
      },
      "ResourceReference" : {
        "type" : "object",
        "properties" : {
          "labels" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "id" : {
            "type" : "string"
          },
          "location" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "mimeType" : {
            "type" : "string"
          },
          "params" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Param"
            }
          },
          "configurationURI" : {
            "type" : "string"
          },
          "secretsURI" : {
            "type" : "string"
          },
          "namespace" : {
            "type" : "string"
          }
        }
      },
...
      "ToolPayload" : {
        "type" : "object",
        "properties" : {
          "toolReference" : {
            "$ref" : "#/components/schemas/ToolReference"
          },
          "configurationData" : {
            "type" : "string"
          },
          "secretsData" : {
            "type" : "string"
          },
          "payload" : {
            "$ref" : "#/components/schemas/ToolReference"
          }
        }
      },
      "ToolReference" : {
        "type" : "object",
        "properties" : {
          "labels" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "uri" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "inputSchema" : {
            "$ref" : "#/components/schemas/InputSchema"
          },
          "namespace" : {
            "type" : "string"
          },
          "configurationURI" : {
            "type" : "string"
          },
          "secretsURI" : {
            "type" : "string"
          }
        }
      },
      "WanakuError" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        }
      },
      "WanakuResponse" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : { }
        }
      },
...
      "WanakuResponseListResourceReference" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ResourceReference"
            }
          }
        }
      },
...
      "WanakuResponseListToolReference" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ToolReference"
            }
          }
        }
      },
...
      "WanakuResponseResourceReference" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "$ref" : "#/components/schemas/ResourceReference"
          }
        }
      },
...
      "WanakuResponseToolReference" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "$ref" : "#/components/schemas/ToolReference"
          }
        }
      }
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "openIdConnect",
        "openIdConnectUrl" : "http://localhost:8543/realms/wanaku/.well-known/openid-configuration",
        "description" : "Authentication"
      }
    }
  },
  "paths" : {
...
    "/api/v1/resources" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseListResourceReference"
                }
              }
            }
          },
          "500" : {
            "description" : "Wanaku error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponse"
                }
              }
            }
          }
        },
        "summary" : "List",
        "tags" : [ "Resources Resource" ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResourceReference"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseResourceReference"
                }
              }
            }
          },
          "500" : {
            "description" : "Wanaku error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Expose",
        "tags" : [ "Resources Resource" ]
      }
    },
    "/api/v1/resources/with-payload" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResourcePayload"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseResourceReference"
                }
              }
            }
          },
          "500" : {
            "description" : "Wanaku error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Expose With Payload",
        "tags" : [ "Resources Resource" ]
      }
    },
    "/api/v1/resources/{resourceName}" : {
      "put" : {
        "parameters" : [ {
          "name" : "resourceName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResourceReference"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Wanaku error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Update",
        "tags" : [ "Resources Resource" ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "resourceName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Wanaku error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponse"
                }
              }
            }
          }
        },
        "summary" : "Remove",
        "tags" : [ "Resources Resource" ]
      }
    },
    "/api/v1/tools" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseListToolReference"
                }
              }
            }
          },
          "500" : {
            "description" : "Wanaku error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponse"
                }
              }
            }
          }
        },
        "summary" : "List",
        "tags" : [ "Tools Resource" ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ToolReference"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseToolReference"
                }
              }
            }
          },
          "500" : {
            "description" : "Wanaku error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Add",
        "tags" : [ "Tools Resource" ]
      }
    },
    "/api/v1/tools/with-payload" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ToolPayload"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseToolReference"
                }
              }
            }
          },
          "500" : {
            "description" : "Wanaku error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Add With Payload",
        "tags" : [ "Tools Resource" ]
      }
    },
    "/api/v1/tools/{toolName}" : {
      "put" : {
        "parameters" : [ {
          "name" : "toolName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ToolReference"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Wanaku error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Update",
        "tags" : [ "Tools Resource" ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "toolName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Wanaku error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponse"
                }
              }
            }
          }
        },
        "summary" : "Remove",
        "tags" : [ "Tools Resource" ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "toolName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseToolReference"
                }
              }
            }
          },
          "500" : {
            "description" : "Wanaku error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponse"
                }
              }
            }
          }
        },
        "summary" : "Get By Name",
        "tags" : [ "Tools Resource" ]
      }
    }
  },
  "info" : {
    "title" : "wanaku-router-backend API",
    "version" : "0.0.9-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}